only creator can delete item or change if vegan or not

+++ application/items/views.py
@@ -5,10 +5,12 @@ from application import app, db, login_required, login_manager
 from application.items.models import Item
 from application.items.forms import ItemForm
 
+
 @app.route("/items/new/")
 @login_required(role="ANY")
 def items_form():
-    return render_template("items/new.html", form = ItemForm())
+    return render_template("items/new.html", form=ItemForm())
+
 
 @app.route("/items/", methods=["POST"])
 @login_required(role="ADMIN")
@@ -16,8 +18,8 @@ def items_create():
     form = ItemForm(request.form)
 
     if not form.validate():
-        return render_template("items/new.html", form = form)
-    
+        return render_template("items/new.html", form=form)
+
     i = Item(form.name.data)
     i.vegan = form.vegan.data
     i.account_id = current_user.id
@@ -27,26 +29,36 @@ def items_create():
 
     return redirect(url_for("items_index"))
 
+
 @app.route("/items/", methods=["GET"])
 def items_index():
-    return render_template("items/list.html", items = Item.query.all())
+    return render_template("items/list.html", items=Item.query.all())
 
-@app.route("/items/<item_id>/", methods = ["POST"])
+
+@app.route("/items/<item_id>/", methods=["POST"])
 @login_required(role="ANY")
-def items_set_vegan(item_id): # VALITSE JOKU MUU KUIN DONE, MUOKKAA MODELS.PY
+def items_set_vegan(item_id):  # VALITSE JOKU MUU KUIN DONE, MUOKKAA MODELS.PY
     i = Item.query.get(item_id)
     if i.account_id != current_user.id:
-        return login_manager.unauthorized()
+        return redirect(url_for("items_index"))
 
-    i.vegan = True
+    if (i.vegan == False):
+        i.vegan = True
+
+    else:
+        i.vegan = False
     db.session().commit()
     return redirect(url_for("items_index"))
 
-@app.route("/items/<item_id>/delete", methods = ["POST"])
+
+@app.route("/items/<item_id>/delete", methods=["POST"])
 @login_required(role="ANY")
 def item_delete(item_id):
     i = Item.query.get(item_id)
+
+    if i.account_id != current_user.id:
+        return redirect(url_for("items_index"))
+        
     db.session.delete(i)
     db.session.commit()
     return redirect(url_for("items_index"))
-
